requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
youtube-transcript-api==0.6.2
openai==1.3.0
pydantic==2.5.0
jinja2==3.1.2
aiofiles==23.2.1
isodate==0.6.1

